+++ includes/class-migration.php
@@ -126,9 +126,8 @@ class Migration {
 				$old_name              = $post->post_name;
 				$old_prefix_normalized = Manager::normalize_slug( $old_prefix );
 				if ( strpos( $old_name, $old_prefix_normalized ) === 0 ) {
-					$page_id_normalized = substr( $old_name, strlen( $old_prefix_normalized ) );
-					$page_id            = Manager::denormalize_slug( $page_id_normalized );
-					$new_name           = Manager::normalize_slug( $new_prefix . $page_id );
+					$page_id  = substr( $old_name, strlen( $old_prefix_normalized ) );
+					$new_name = Manager::normalize_slug( $new_prefix . $page_id );
 
 					$result = wp_update_post(
 						array(
@@ -174,8 +173,7 @@ class Migration {
 				$prefix_normalized = Manager::normalize_slug( $new_prefix );
 
 				if ( strpos( $post_name, $prefix_normalized ) === 0 ) {
-					$page_id_normalized = substr( $post_name, strlen( $prefix_normalized ) );
-					$page_id            = Manager::denormalize_slug( $page_id_normalized );
+					$page_id = substr( $post_name, strlen( $prefix_normalized ) );
 
 					if ( isset( $capability_map[ $page_id ] ) ) {
 						$new_capability = $capability_map[ $page_id ];