+++ includes/metabox/actions.php
@@ -4,30 +4,15 @@
  *
  * @package CodeSoup\Options
  *
- * @var \WP_Post $post Post object.
+ * @var \WP_Post                     $post Post object.
+ * @var array                        $args Metabox arguments.
+ * @var \CodeSoup\Options\Manager    $args['manager'] Manager instance.
  */
 
 // Don't allow direct access to file.
-defined( 'ABSPATH' ) || die;
-?>
+defined( 'ABSPATH' ) || die; ?>
+
 <div class="submitbox" id="submitpost">
-	<div id="delete-action">
-		<?php
-		if ( current_user_can( 'delete_post', $post->ID ) ) {
-			// phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedVariableFound
-			$delete_url = get_delete_post_link(
-				$post->ID,
-				'',
-				true
-			);
-			printf(
-				'<a class="submitdelete deletion" href="%s">%s</a>',
-				esc_url( $delete_url ),
-				esc_html__( 'Delete', 'codesoup-options' )
-			);
-		}
-		?>
-	</div>
 	<div id="publishing-action">
 		<span class="spinner"></span>
 		<?php